fix(#1716): Remove obsolete data type transformer implementations
- Data type transformer implementations have been moved to the Camel core project
- Using those transformers in Camel works with version 4.2.0 onwards
