+++ kamelets/protobuf-serialize-action.kamelet.yaml
@@ -46,7 +46,7 @@ spec:
   template:
     beans:
       - name: schemaResolver
-        type: "#class:org.apache.camel.kamelets.utils.format.converter.protobuf.ProtobufSchemaResolver"
+        type: "#class:org.apache.camel.kamelets.utils.format.schema.protobuf.ProtobufSchemaResolver"
         property:
           - key: schema
             value: '{{schema:}}'
