+++ library/camel-kamelets-utils/pom.xml
@@ -71,35 +71,10 @@
             <artifactId>camel-kafka</artifactId>
         </dependency>
 
-        <!-- Optional dependencies for data type conversion -->
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-aws2-ddb</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-aws2-s3</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-aws2-sqs</artifactId>
-            <scope>provided</scope>
-        </dependency>
+        <!-- Dependencies for azure credential configuration -->
         <dependency>
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-azure-storage-blob</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-google-storage</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel</groupId>
-            <artifactId>camel-google-sheets</artifactId>
             <scope>provided</scope>
         </dependency>
