+++ library/camel-kamelets-utils/src/main/java/org/apache/camel/kamelets/utils/format/schema/DelegatingSchemaResolver.java
@@ -19,10 +19,10 @@ package org.apache.camel.kamelets.utils.format.schema;
 
 import org.apache.camel.Exchange;
 import org.apache.camel.Processor;
+import org.apache.camel.component.jackson.avro.transform.AvroSchemaResolver;
+import org.apache.camel.component.jackson.transform.JsonSchemaResolver;
 import org.apache.camel.kamelets.utils.format.MimeType;
-import org.apache.camel.kamelets.utils.format.converter.avro.AvroSchemaResolver;
-import org.apache.camel.kamelets.utils.format.converter.json.JsonSchemaResolver;
-import org.apache.camel.kamelets.utils.format.converter.protobuf.ProtobufSchemaResolver;
+import org.apache.camel.kamelets.utils.format.schema.protobuf.ProtobufSchemaResolver;
 import org.apache.camel.util.ObjectHelper;
 
 /**
