+++ library/camel-kamelets/src/main/resources/kamelets/data-type-action.kamelet.yaml
@@ -45,7 +45,6 @@ spec:
         description: Defines the data type that will be applied by this action. Apache Camel and the Kamelet catalog support different data types and performs automatic message conversion according to the given type.
         type: string
   dependencies:
-    - "mvn:org.apache.camel.kamelets:camel-kamelets-utils:4.2.0-SNAPSHOT"
     - "camel:kamelet"
     - "camel:core"
   template:
