+++ pom.xml
@@ -60,7 +60,7 @@
         <apache-rat-plugin.version>0.15</apache-rat-plugin.version>
         <cyclonedx-maven-plugin-version>2.7.10</cyclonedx-maven-plugin-version>
 
-        <camel.version>4.1.0</camel.version>
+        <camel.version>4.2.0</camel.version>
         <camel.k.crds.version>2.0.1</camel.k.crds.version>
 
         <!-- Versions used inside Kamelets (add them also to the dependencyManagement section and the groovy script below) -->
@@ -318,7 +318,7 @@
                         "version.io.apicurio.apicurio-registry-serdes-avro-serde": "${version.io.apicurio.apicurio-registry-serdes-avro-serde}",
                         "version.com.azure.azure-identity": "${version.com.azure.azure-identity}",
                         "version.com.azure.azure-data-schemaregistry-apacheavro": "${version.com.azure.azure-data-schemaregistry-apacheavro}",
-                        "version.com.microsoft.azure.azure-schemaregistry-kafka-avro": "${version.com.microsoft.azure.azure-schemaregistry-kafka-avro}"                        
+                        "version.com.microsoft.azure.azure-schemaregistry-kafka-avro": "${version.com.microsoft.azure.azure-schemaregistry-kafka-avro}"
                     ])
                     </source>
                 </configuration>
